+++ src/app/components/info/info-item/info-item.component.ts
@@ -8,11 +8,12 @@ import { LocationModalComponent } from '../../location/location-modal/location-m
 import { GalleryModalComponent } from '../../gallery/gallery-modal/gallery-modal.component';
 import { ImagesService, Netlifile } from 'src/app/services/images.service';
 import { map, takeUntil } from 'rxjs/operators';
-import { Press, PressService } from 'src/app/services/press.service';
+import { PressService } from 'src/app/services/press.service';
 import { TicketModalComponent } from '../../ticket/ticket-modal/ticket-modal.component';
 import { AboutModalComponent } from '../../../modules/about/about-modal/about-modal.component';
-import { Job, StaffService } from '../../../services/staff.service';
+import { StaffService } from '../../../services/staff.service';
 import { environment } from '../../../../environments/environment';
+import { Job } from '../../../models/staff.models';
 
 @Component({
   selector: 'app-info-item',
