+++ src/app/models/staff.models.ts
@@ -1,0 +1,14 @@
+export interface Staff {
+	imageSrc: string;
+	bio: string;
+	name: string;
+	topic: string;
+	active: boolean;
+	sortOrder: number;
+}
+
+export interface Job {
+	isJobSharing: boolean;
+	name: string;
+	values: string[];
+}
